requirements.txt
docling==2.63.0
python-dotenv==1.0.1
requests==2.32.3
PyInstaller==6.11.1

# OCR движки
# EasyOCR - простая установка, работает из коробки (рекомендуется для Windows)
easyocr==1.7.2

# Tesserocr - быстрее, но сложная установка на Windows
# Для установки на Windows:
# 1. Скачайте .whl файл: https://github.com/simonflueckiger/tesserocr-windows_build/releases
# 2. Установите: pip install tesserocr-2.x.x-cpXXX-cpXXX-win_amd64.whl
# На Linux/Mac: pip install tesserocr
# tesserocr>=2.6.0
